requirements.txt
numpy==1.17.0
opencv-python
torch==1.4.0
torchvision==0.5.0
matplotlib
tqdm
tb-nightly
future
Pillow
enforce
onnx==1.6.0
onnxruntime>1.0.0
mypy
autopep8
pyyaml
easydict
pandas
comet_ml    
multipledispatch 
dvc[gs]
optuna
anytree
scipy
albumentations
pytest
pytest-cov
sklearn
flatten-dict
graphviz
seaborn
psutil
gpustat
thop
setuptools>=41.0.0